bug fixes in drqa
The default key for the NULL token is `'__NULL__'` (See line 59 in
parlai/core/dict.py), but `'<NULL>'` is used in line 193 and line 226
of parlai/agents/drqa/drqa.py. As a result, `'<NULL>'` is treated as an
unknown word and the index of `'__UNK__'` is actually used for padding,
which is not expected.

The string 'merge_mode' only appears in line 80 of
parlai/agents/drqa/rnn\_reader.py, which seems to be a typo. It should
be 'question\_mode'.