Update google_sgd to use slot metrics
Tested with a pretrained model on taskmaster2 + validated that the response made sense.

Note that I did have to comment out the "resp.startswith" bit in the slot parsing to use the call token from taskmaster2 since the model expected a different token. The values there were all, unsurprisingly, zeroes, but this seemed fine for the purposes of validation.