Modernize natural_questions with add_cmdline_args; add flag for context (#3440)
* Modernize natural_questions with add_cmdline_args; add flag for context

I want to get rid of the context and figured I should add a new cmdline arg. Natural questions doesn't currently use cmdline arguments, so update it so that it does.

Keep the old versions around for backward compatability purposes.

Test Plan:

Variations of the flag with

```
parlai display_data -t natural_questions -dt valid:stream
```

ensuring that the output is reasonable.

* Remove unused teachers

Didn't want to remove them outright in case someone was using them, but search suggests they're not. (Which is consistent with @mojtaba-komeili's comment prior)